+++ common/prompt.ts
@@ -1,12 +1,11 @@
 import type { GenerateRequestV2 } from '../srv/adapter/type'
-import type { AppSchema } from './types/schema'
+import type { AppSchema, TokenCounter } from './types'
 import { AIAdapter, NOVEL_MODELS, OPENAI_MODELS, ThirdPartyFormat } from './adapters'
 import { formatCharacter } from './characters'
 import { defaultTemplate } from './mode-templates'
 import { buildMemoryPrompt } from './memory'
 import { defaultPresets, getFallbackPreset, isDefaultPreset } from './presets'
 import { parseTemplate } from './template-parser'
-import { TokenCounter } from './tokenize'
 import { getBotName, trimSentence } from './util'
 import { Memory } from './types'
 import { promptOrderToTemplate } from './prompt-order'
