+++ common/tokenize.ts
@@ -1,17 +1,5 @@
+// @ts-ignore
 import { embedApi } from '/web/store/embeddings'
-
-export type TokenCounter = (text: string) => number | Promise<number>
-
-export type Encoder = {
-  count: (text: string) => number
-  encode: (text: string) => number[]
-  decode: (tokens: number[]) => string
-}
-
-export type Tokenizer = {
-  decode: (tokens: Int32Array) => string
-  encode: (value: string) => Int32Array
-}
 
 export async function encode(text: string) {
   return embedApi.encode(text)
@@ -27,5 +15,5 @@ export async function tokenize(text: string) {
 }
 
 export async function getEncoder() {
-  return (text: string) => embedApi.encode(text).then((res) => res.length)
+  return (text: string) => embedApi.encode(text).then((res: number[]) => res.length)
 }
