+++ common/types/index.ts
@@ -6,3 +6,16 @@ export * from './schema'
 export * from './texttospeech-schema'
 
 export { UI, Sprite, Memory }
+
+export type TokenCounter = (text: string) => number | Promise<number>
+
+export type Encoder = {
+  count: (text: string) => number
+  encode: (text: string) => number[]
+  decode: (tokens: number[]) => string
+}
+
+export type Tokenizer = {
+  decode: (tokens: Int32Array) => string
+  encode: (value: string) => Int32Array
+}
