+++ srv/tokenize.ts
@@ -6,9 +6,9 @@ import { encoding_for_model } from '@dqbd/tiktoken'
 import { AIAdapter, NOVEL_MODELS, OPENAI_MODELS } from '../common/adapters'
 import gpt from 'gpt-3-encoder'
 import { resolve } from 'path'
-import { Encoder, TokenCounter, Tokenizer } from '../common/tokenize'
 import * as nai from 'nai-js-tokenizer'
 import { logger } from './logger'
+import { Encoder, TokenCounter, Tokenizer } from '/common/types'
 
 const claudeJson = readFileSync(resolve(__dirname, 'sp-models', 'claude.json'))
 const pileJson = readFileSync(resolve(__dirname, 'sp-models', 'pile_tokenizer.json'))
